requirements.txt
disnake==2.9.0
owoify_py==2.0.0
python-dotenv==1.0.0
Requests==2.31.0
